Consolidate py reverse http uri code into a mixin
