+++ lib/msf/core/payload/python/reverse_tcp.rb
@@ -63,4 +63,3 @@ module Payload::Python::ReverseTcp
 end
 
 end
-
